Sophie Chaveli BlackJack Game submission
PlayBlackJack is the main game. Running it will start a game of single player BlackJack. Card is a supporting class.

AllTests is a JUnit test case suite. To run all of the tests, runTestSuite can be run and it will run every unit test that will show that the scenarios given are satisfied.